tulos

+++ configure.sh
@@ -1,29 +1,42 @@
 set -e
 
+function install_aur_package {
+    if (pacman -Q package-query >/dev/null) then
+        yaourt -S $1 --noconfirm
+    fi
+}
+
 sudo pacman -Syu - < packages --noconfirm
 
-cd ~
-rm -Rf package-query
-git clone https://aur.archlinux.org/package-query.git
-cd package-query
-makepkg -si --noconfirm
-rm -Rf package-query
+if (pacman -Q package-query >/dev/null) then
+    cd ~
+    rm -Rf package-query
+    git clone https://aur.archlinux.org/package-query.git
+    cd package-query
+    makepkg -si --noconfirm
+    rm -Rf package-query
+fi
 
-cd ~
-rm -Rf yaourt
-git clone https://aur.archlinux.org/yaourt.git
-cd yaourt
-makepkg -si --noconfirm
-cd ..
-rm -Rf yaourt
+if (pacman -Q yaourt >/dev/null) then
+    cd ~
+    rm -Rf yaourt
+    git clone https://aur.archlinux.org/yaourt.git
+    cd yaourt
+    makepkg -si --noconfirm
+    cd ..
+    rm -Rf yaourt
+fi
 
-cd TulOS
-yaourt -S i3blocks ttf-inconsolata-g ttf-font-awesome --noconfirm
+install_aur_package i3blocks
+install_aur_package ttf-inconsolata-g
+install_aur_package ttf-font-awesome
+install_aur_package xnviewmp
 
 sudo pip install virtualenvwrapper
 
 sh -c "$(wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
 
+cd TulOS
 rm ~/.config/i3/config
 rm ~/.config/i3blocks/config
 rm ~/.config/termite/config
